style(modules/users): move controllers to `/controllers`
