+++ src/modules/users/controllers/index.ts
@@ -1,0 +1,3 @@
+export * from './users.controller'
+export * from './users-profile.controller'
+export * from './users-playback.controller'
